Add shared install helpers and split mise out of shell setup.
Centralize packager installs via installPkg/installFlatpak/installGroup, improve moss and other distro coverage, always install eza and fastfetch in shell setup, and move mise into a dedicated Development entry shared by language scripts.

+++ core/tabs/linux/applications-setup/browsers/google-chrome.sh
@@ -3,38 +3,39 @@
 . ../../common-script.sh
 
 installChrome() {
-  if ! command_exists google-chrome; then
-    printf "%b\n" "${YELLOW}Installing Google Chrome...${RC}"
-    case "$PACKAGER" in
-    apt-get | nala)
-      curl -O https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-      "$ESCALATION_TOOL" "$PACKAGER" install -y ./google-chrome-stable_current_amd64.deb
-      ;;
-    zypper)
-      "$ESCALATION_TOOL" "$PACKAGER" addrepo http://dl.google.com/linux/chrome/rpm/stable/x86_64 Google-Chrome
-      "$ESCALATION_TOOL" "$PACKAGER" refresh
-      "$ESCALATION_TOOL" "$PACKAGER" --non-interactive install google-chrome-stable
-      ;;
-    pacman)
-      "$AUR_HELPER" -S --needed --noconfirm google-chrome
-      ;;
-    dnf)
-      "$ESCALATION_TOOL" "$PACKAGER" install -y fedora-workstation-repositories
-      "$ESCALATION_TOOL" "$PACKAGER" config-manager --set-enabled google-chrome
-      "$ESCALATION_TOOL" "$PACKAGER" install -y google-chrome-stable
-      ;;
-    *)
-      printf "%b\n" "${RED}Unsupported package manager: ""$PACKAGER""${RC}"
-      exit 1
-      ;;
-    esac
-  else
-    printf "%b\n" "${GREEN}Google Chrome Browser is already installed.${RC}"
-  fi
+    if ! command_exists google-chrome && ! command_exists com.google.Chrome; then
+        printf "%b\n" "${YELLOW}Installing Google Chrome...${RC}"
+        case "$PACKAGER" in
+            apt-get|nala)
+                curl -O https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+                installPkg ./google-chrome-stable_current_amd64.deb
+                ;;
+            zypper)
+                "$ESCALATION_TOOL" "$PACKAGER" addrepo http://dl.google.com/linux/chrome/rpm/stable/x86_64 Google-Chrome
+                "$ESCALATION_TOOL" "$PACKAGER" refresh
+                installPkg google-chrome-stable
+                ;;
+            pacman)
+                installAurPkg google-chrome
+                ;;
+            dnf)
+                installPkg fedora-workstation-repositories
+                "$ESCALATION_TOOL" "$PACKAGER" config-manager --set-enabled google-chrome
+                installPkg google-chrome-stable
+                ;;
+            moss|rpm-ostree|apk|xbps-install|eopkg|pkg)
+                installFlatpak com.google.Chrome
+                ;;
+            *)
+                unsupportedPackager
+                ;;
+        esac
+    else
+        printf "%b\n" "${GREEN}Google Chrome Browser is already installed.${RC}"
+    fi
 }
 
 checkEnv
 checkEscalationTool
 checkAURHelper
 installChrome
-
